MAGNOLIA-3456. Backport of r41203. Ensure activation doesn't fail when content for deletion is already deleted.
git-svn-id: file:///home/migrator/subversion-migration/svn-repo-copy/community/magnolia/trunk@41205 a15fbcf9-bee4-0310-bafe-d02d26ad3b5f

+++ magnolia-module-exchange-simple/src/main/java/info/magnolia/module/exchangesimple/BaseSyndicatorImpl.java
@@ -307,13 +307,19 @@ public abstract class BaseSyndicatorImpl implements Syndicator {
                 this.activate(subscriber, activationContent, path);
             }
             if (Boolean.parseBoolean(activationContent.getproperty(ItemType.DELETED_NODE_MIXIN))) {
-                if (content instanceof ContentVersion) {
-                    // replace versioned content with the real node
-                    content = content.getHierarchyManager().getContentByUUID(content.getUUID());
+                final HierarchyManager hm = content.getHierarchyManager();
+                String uuid = content.getUUID();
+                if (StringUtils.isNotBlank(uuid)) {
+                    if (content instanceof ContentVersion) {
+                        // replace versioned content with the real node
+                        content = hm.getContentByUUID(uuid);
+                    }
+                    Content parentContent = content.getParent();
+                    content.delete();
+                    parentContent.save();
+                } else {
+                    log.warn("Content {}:{} was already removed.", new String[] {hm.getName(), path});
                 }
-                Content parentContent = content.getParent();
-                content.delete();
-                parentContent.save();
             } else {
                 this.updateActivationDetails(path);
             }
